requirements.txt
absl-py==2.0.0
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beanie==1.26.0
beautifulsoup4==4.12.2
binarytree==6.5.1
bleach==6.1.0
cachetools==4.2.4
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
comm==0.2.1
contourpy==1.1.0
customtkinter==5.1.3
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.27.2
distlib==0.3.7
Django==4.2.4
dnspython==2.6.1
email_validator==2.1.2
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastcore==1.5.29
fastjsonschema==2.19.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.5.4
google==3.0.0
google-api-core==1.34.0
google-auth-oauthlib==1.2.0
google-cloud-language==2.10.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
grpcio==1.56.0
grpcio-status==1.48.2
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.4
importlib_metadata==7.1.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
keras==2.15.0
kiwisolver==1.4.4
lazy-model==0.2.0
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.9
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.1.0
motor==3.4.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.1.2
notebook_shim==0.2.3
np==1.0.2
numpy==1.25.0
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.5
overrides==7.6.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
platformdirs==3.10.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo @ file:///private/tmp/py3cairo-20240213-5896-uig1tg/pycairo-1.26.0
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
pyenchant==3.2.2
pygame==2.5.2
Pygments==2.17.2
pymongo==4.7.3
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.2
setuptools-scm==8.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.39.1
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.10.0
uri-template==1.3.0
urllib3==2.0.3
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.24.3
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.14.1
zipp==3.18.1
